+++ R/crew_controller_local.R
@@ -30,7 +30,6 @@ crew_controller_local <- function(
   seconds_wall = Inf,
   seconds_exit = NULL,
   retry_tasks = TRUE,
-  log_resources = NULL,
   tasks_max = Inf,
   tasks_timers = 0L,
   reset_globals = TRUE,
@@ -39,8 +38,10 @@ crew_controller_local <- function(
   garbage_collection = FALSE,
   launch_max = 5L,
   r_arguments = c("--no-save", "--no-restore"),
+  options_metrics = crew::crew_options_metrics(),
+  options_local = crew::crew_options_local(),
   local_log_directory = NULL,
-  local_log_join = TRUE
+  local_log_join = NULL
 ) {
   crew_deprecate(
     name = "seconds_exit",
@@ -50,6 +51,26 @@ crew_controller_local <- function(
     condition = "warning",
     value = seconds_exit
   )
+  crew_deprecate(
+    name = "local_log_directory",
+    date = "2024-10-8",
+    version = "0.9.5.9012",
+    alternative = "options_local argument",
+    condition = "warning",
+    value = local_log_directory
+  )
+  crew_deprecate(
+    name = "local_log_join",
+    date = "2024-10-8",
+    version = "0.9.5.9012",
+    alternative = "options_local argument",
+    condition = "warning",
+    value = local_log_join
+  )
+  options_local$log_directory <- local_log_directory %|||%
+    options_local$log_directory
+  options_local$log_join <- local_log_join %|||%
+    options_local$log_join
   client <- crew_client(
     name = name,
     workers = workers,
@@ -60,8 +81,7 @@ crew_controller_local <- function(
     tls_config = tls_config,
     seconds_interval = seconds_interval,
     seconds_timeout = seconds_timeout,
-    retry_tasks = retry_tasks,
-    log_resources = log_resources
+    retry_tasks = retry_tasks
   )
   launcher <- crew_launcher_local(
     name = name,
@@ -79,6 +99,8 @@ crew_controller_local <- function(
     launch_max = launch_max,
     tls = tls,
     r_arguments = r_arguments,
+    options_metrics = options_metrics,
+    options_local = options_local,
     local_log_directory = local_log_directory,
     local_log_join = local_log_join
   )