+++ R/crew_package.R
@@ -18,6 +18,7 @@
 #'   [`rrq`](https://mrc-ide.github.io/rrq/),
 #'   [`clustermq`](https://mschubert.github.io/clustermq/),
 #'   and [`batchtools`](https://mllg.github.io/batchtools/).
+#' @importFrom autometric log_start log_stop
 #' @importFrom cli cli_progress_bar cli_progress_done cli_progress_update
 #' @importFrom data.table rbindlist
 #' @importFrom getip getip
@@ -28,7 +29,7 @@
 #'   unresolved until_ tls_config
 #' @importFrom processx process
 #' @importFrom promises promise
-#' @importFrom ps ps_cmdline ps_handle ps_memory_info ps_pid ps_is_running
+#' @importFrom ps ps_cmdline ps_handle ps_pid ps_is_running
 #'   ps_status ps_username
 #' @importFrom R6 R6Class
 #' @importFrom rlang abort as_function enquo is_installed is_named quo_squash